[RFR] Tag tier 1 test

+++ cypress/integration/tests/administrator/credentials/remove_creds_app.test.ts
@@ -46,6 +46,9 @@ describe("Validation of Source Control Credentials", { tags: "@tier1" }, () => {
     });
 
     after("Cleanup", () => {
+        // Prevent hook from running, if the tag is excluded from run
+        if (hasToBeSkipped("@tier1")) return;
+
         application.delete();
     });
 });
